Load function for loading runfile
